requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
block-disposable-email==2.0.0
celery==5.4.0
certifi==2025.1.31
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.7
django-admin-interface==0.29.4
django-ckeditor==6.7.2
django-colorfield==0.12.0
django-cors-headers==4.7.0
django-disposable-email-checker==2.0.4
django-js-asset==3.1.2
django-model-utils==5.0.0
django-ratelimit==4.1.0
django-redis==5.4.0
django-taggit==6.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
freezegun==1.5.1
hiredis==3.1.0
idna==3.10
kombu==5.5.1
msgpack==1.1.0
pillow==11.1.0
prompt_toolkit==3.0.50
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
